+++ R/fit.R
@@ -50,7 +50,6 @@ cpm_spec <- function(
   na_action <- match.arg(na_action)
 
   new_cpm_spec(
-    call = match.call(),
     params = list(
       thresh_method = thresh_method,
       thresh_level = thresh_level,
@@ -96,32 +95,80 @@ fit.cpm_spec <- function(
   conmat,
   behav,
   ...,
-  covariates = NULL,
-  confounds = NULL
+  covariates = NULL
 ) {
   call <- match.call()
   call[[1]] <- quote(fit)
-
   fit_cpm_workflow(
     call = call,
     object = object,
     conmat = conmat,
     behav = behav,
-    covariates = covariates,
-    confounds = confounds
+    covariates = covariates
   )
 }
 
-new_cpm_spec <- function(call, params) {
+new_cpm_spec <- function(params) {
   structure(
-    list(
-      call = call,
-      params = params
-    ),
+    list(params = params),
     class = "cpm_spec"
   )
 }
 
 `%||%` <- function(x, y) {
   if (is.null(x)) y else x
+}
+
+fit_cpm_workflow <- function(
+  call,
+  object,
+  conmat,
+  behav,
+  covariates
+) {
+  params <- object$params
+
+  normalized <- normalize_inputs(conmat, behav, covariates)
+  behav <- normalized$behav
+  covariates <- normalized$covariates
+
+  include_cases <- resolve_include_cases(
+    conmat,
+    behav,
+    covariates,
+    params$na_action
+  )
+
+  kfolds <- resolve_kfolds(params$kfolds, include_cases)
+  folds <- crossv_kfold(include_cases, kfolds)
+  edges <- init_edges(params$return_edges, conmat, kfolds)
+  pred <- init_pred(behav)
+  cv_result <- fit_predict_cv(
+    conmat,
+    behav,
+    covariates,
+    include_cases,
+    folds,
+    params$thresh_method,
+    params$thresh_level,
+    params$bias_correct,
+    params$return_edges,
+    pred,
+    edges
+  )
+
+  new_cpm(
+    call = call,
+    folds = folds,
+    behav = cv_result$real,
+    pred = cv_result$pred,
+    edges = cv_result$edges,
+    params = list(
+      covariates = !is.null(covariates),
+      thresh_method = params$thresh_method,
+      thresh_level = params$thresh_level,
+      kfolds = kfolds,
+      bias_correct = params$bias_correct
+    )
+  )
 }